requirements.txt
asgiref==3.3.4
boto3==1.17.84
botocore==1.20.84
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.3.1
djangorestframework==3.12.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
jmespath==0.10.0
numpy==1.20.3
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
